Build: Disable non-core source code compilation
The incoming refactoring changes will break compilation of all
components, so in order to aim low and achieve it all compile only core.

PS: Refactoring the whole project structure in a continual fine-grained
commits seems a bit hard to me, but I am trying :). This patch seems
a non-perfect but reasonable solution.