Refactor: Put a unified CMakeLists.txt file in `build/`
Remove unnecessary complication with many different `CMakeLists.txt`
files by putting the build recipes inside a single one. Note that it is
not contained inside the root of the project but inside `build/` instead.

I am not really sure about that but for now, as long as the library
implementation stays at such minimal volume, I think this way it would
result in a cleaner project structure.