Refactor: Integrate benchmarks into new project structure
Previously benchmarks were considered as totally separate entities to
other kinds of tests and now I have unified them a bit. They are still
placed in a logically different location but rather than reside right
next to the core library, I have put them in tests and are now labeled
as *some form of testing mechanism* in my head.

I think that's the proper way to think of them :).

Also, they are toggled with the same flag that toggles building all
other tests. However, the μbenchmarks are invoked using the `bench`
command and not the `test` one.

Test plan:
```
    $ ./build/abc.sh clean
    $ ./build/abc.sh build
    [...]
    -- Telamon: Benchmarks: Add LockFreeSampleBench μbenchmark
    -- Telamon: Benchmarks: Add WaitFreeSampleBench μbenchmark
    [...]

    $ ./build/abc.sh bench
    Invoking μbenchmark 'build/out/bin/bench_LockFreeSampleBench' ...
    2022-11-29T20:04:07+02:00
    Running ./build/out/bin/bench_LockFreeSampleBench
    Run on (4 X 2933 MHz CPU s)
    CPU Caches:
      L1 Data 32 KiB (x2)
      L1 Instruction 32 KiB (x2)
    [...]
```